test(runtime): isolate basic control-flow regressions
